requirements.txt
python-dotenv
pymupdf
pandas
spacy
pdfplumber
openai
requests


#install via: pip install -r requirements.txt

